Rename wrapper/ to wrapper_backend/
Carves out room for a sibling wrapper-frontend/ Svelte project.
Python can't have hyphens in package names, so the backend keeps
the underscore; the frontend will use the npm-conventional hyphen.

Updates imports, pyproject.toml, .pre-commit-config.yaml,
.gitignore, start_wrapper.sh, logger names, and all doc
references. Also corrects a stale README claim that
default_lines: was removed (it was restored as
optional_field_lines:).
