+++ .pre-commit-config.yaml
@@ -4,9 +4,9 @@ repos:
     hooks:
       - id: ruff
         args: [--fix]
-        files: ^wrapper/
+        files: ^wrapper_backend/
       - id: ruff-format
-        files: ^wrapper/
+        files: ^wrapper_backend/
   - repo: local
     hooks:
       - id: mypy
@@ -14,5 +14,5 @@ repos:
         entry: uv run mypy
         language: system
         types: [python]
-        files: ^wrapper/
+        files: ^wrapper_backend/
         pass_filenames: false
