+++ start_wrapper.sh
@@ -1,4 +1,4 @@
 #!/usr/bin/env bash
 set -euo pipefail
 cd "$(dirname "$0")"
-exec uv run python -m wrapper "${1:-geometry-divB.yml}" "${@:2}"
+exec uv run python -m wrapper_backend "${1:-geometry-divB.yml}" "${@:2}"
